requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.10.1
djangorestframework==3.13.1
Pillow==8.4.0
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
